Transfer hooks to new module (#2)
