+++ src/cleantest/__init__.py
@@ -2,4 +2,4 @@
 # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-from .config import Configure, StartEnvHook, StartTestletHook, StopEnvHook, StopTestletHook
+from .control import Configure
